requirements.txt
streamlit
python-dotenv
nest-asyncio

# LangChain
langchain
langchain-community
langchain-groq

# Vector DB
faiss-cpu

# Embeddings
sentence-transformers

# Document loaders
pypdf
docx2txt
